back/requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
debugpy==1.5.1
decorator==5.1.0
Django==4.0
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
djangorestframework==3.13.1
django-six==1.0.4
entrypoints==0.3
ipykernel==6.6.0
drf-yasg==1.20.0
ipython==7.30.1
psycopg2==2.9.2
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
screen==1.0.1
sqlparse==0.4.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
xlwt==1.3.0
Pillow==8.4.0
django-light==0.1.0.post3
django-ckeditor==6.0.0
django-cors-headers==3.11.0
